project/requirements.txt
numpy==1.26.4
pandas==2.2.2
numba==0.59.1
llvmlite==0.42.0
pyyaml==6.0.1
pytest==8.2.0
